Changed: Enable multi-threaded Flask
In single-threaded mode the Flask app would repeatedly hang when
receiving another request while already in the process of e.g.
transferring a range of an audio.

We now connect and close the DB connection on/after every request, since
it can't be shared across threads. The overhead of doing this should
be fine for the purposes of using this as a local server for a single
user. Otherwise another DB engine would have to be used.